replace executeFormulaOrSync with executeThunk. embed error handling into it
